[UnitTests] Remove remaining tests for legacy Kinetics
